refactor(philo): restructure directories and resolve norminette issues

+++ philo/srcs/logging.c
@@ -6,16 +6,16 @@
 /*   By: nburchha <nburchha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/29 01:48:14 by niklasburch       #+#    #+#             */
-/*   Updated: 2024/06/01 14:00:50 by nburchha         ###   ########.fr       */
+/*   Updated: 2025/11/19 15:26:04 by nburchha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/philo.h"
+#include "philo.h"
 
 void	print_status(t_philo *philo, char *status)
 {
 	pthread_mutex_lock(&philo->data->print_mutex);
-	printf("%llu %d %s\n", get_time() - philo->data->start,
+	printf("%llu %d %s\n", (unsigned long long)(get_time() - philo->data->start),
 		philo->id, status);
 	pthread_mutex_unlock(&philo->data->print_mutex);
 }
